+++ crates/vercel_runtime/src/ipc/mod.rs
@@ -1,0 +1,3 @@
+pub mod core;
+pub mod log;
+pub mod metric;
